gfx missing file update
